Move/rename tests to match source file hierarchy
