--- src/__tests__/ssr.spec.ts
@@ -1,7 +1,0 @@
-import flatpickr from "../index";
-
-describe("flatpickr SSR", () => {
-  it("can be imported", () => {
-    expect(typeof flatpickr).toEqual("function");
-  });
-});
